Initial refactor for V2 sheet code and CSS sharing.
