+++ .editorconfig
@@ -1,9 +1,12 @@
 root = true
 
 [*]
+charset = utf-8
+end_of_line = lf
 indent_style = space
 indent_size = 2
 insert_final_newline = true
+trim_trailing_whitespace = true
 
 [*.hbs]
 indent_size = 4
