+++ less/dnd5e.less
@@ -18,6 +18,8 @@
 
 /* V2 Styles */
 @import "v2/apps.less";
+@import "v2/forms.less";
+@import "v2/sheets.less";
 @import "v2/activities.less";
 @import "v2/actors.less";
 @import "v2/character.less";
@@ -30,6 +32,7 @@
 
 /* Dark Mode */
 @import "v2/dark/apps.less";
+@import "v2/dark/forms.less";
 @import "v2/dark/actors.less";
 @import "v2/dark/character.less";
 @import "v2/dark/npc.less";
