+++ less/elements.less
@@ -128,9 +128,10 @@ item-list-controls search, .dnd5e2 .filter-element search {
     height: unset;
     line-height: 1;
     font-family: var(--dnd5e-font-roboto);
+    outline: none;
 
     &:hover, &:focus-visible {
-      outline: 0;
+      outline: none;
       box-shadow: 0 0 12px var(--dnd5e-color-gold);
     }
   }
