+++ less/v2/actors.less
@@ -14,39 +14,6 @@
     border-radius: 5px;
   }
 
-  > header {
-    background: transparent;
-    position: relative;
-    z-index: 1;
-
-    .window-title { visibility: hidden; }
-
-    /* Edit Mode Toggle */
-    > slide-toggle {
-      --slide-toggle-track-color-checked: var(--dnd5e-color-gold);
-      --slide-toggle-track-color-unchecked: var(--dnd5e-color-light-gray);
-      flex: none;
-
-      .slide-toggle-thumb {
-        line-height: 12px;
-
-        &::before {
-          content: "\f0ad";
-          font-family: var(--font-awesome);
-          color: var(--dnd5e-color-light-gray);
-          font-weight: bold;
-          font-size: var(--font-size-9);
-        }
-      }
-
-      &[checked] .slide-toggle-thumb::before {
-        color: var(--dnd5e-color-gold);
-      }
-    }
-  }
-
-  .window-content { margin-top: -30px; }
-
   /* ---------------------------------- */
   /*  Navigation                        */
   /* ---------------------------------- */
@@ -60,17 +27,6 @@
   /* ---------------------------------- */
 
   .sheet-header {
-    input, select {
-      background: rgb(0 0 0 / 45%);
-      border: 1px solid transparent;
-      transition: all 250ms ease;
-
-      &:hover, &:focus {
-        border: 1px solid var(--dnd5e-color-gold);
-        box-shadow: 0 0 6px var(--dnd5e-color-gold);
-      }
-    }
-
     /* Resting & Special Traits */
     .sheet-header-buttons {
       display: flex;
@@ -88,39 +44,6 @@
       /* Tabbed Content */
       .tab-body .tab:not(.active) { display: none; }
     }
-  }
-
-  /* ---------------------------------- */
-  /*  Badges                            */
-  /* ---------------------------------- */
-
-  .badge {
-    background: transparent no-repeat center / contain;
-    font-family: var(--dnd5e-font-roboto);
-    font-weight: bold;
-    color: var(--color-text-light-0);
-    display: grid;
-    place-content: center;
-    font-size: var(--font-size-30);
-
-    .config-button {
-      color: inherit;
-      font-size: inherit;
-    }
-  }
-
-  /* AC */
-  .ac-badge {
-    width: 68px;
-    height: 68px;
-    text-align: center;
-    line-height: 62px;
-    background-image: url("ui/ac-badge.webp");
-    display: block;
-    place-content: unset;
-    font-size: var(--font-size-23);
-    margin-top: -21px;
-    filter: drop-shadow(0 0 6px var(--color-shadow-dark));
   }
 
   /* ---------------------------------- */
@@ -322,35 +245,6 @@
       background: var(--dnd5e-color-scrollbar);
     }
   }
-
-  .prosemirror menu {
-    .pm-dropdown {
-      background: none;
-      &:hover { background: var(--color-hover-bg); }
-    }
-
-    button {
-      display: unset;
-      text-transform: unset;
-      box-shadow: none;
-      font-family: var(--font-primary);
-      font-size: var(--font-size-14);
-      font-weight: normal;
-      line-height: 28px;
-    }
-  }
-
-  /* ---------------------------------- */
-  /*  Child Creation                    */
-  /* ---------------------------------- */
-
-  .create-child {
-    display: none;
-    position: absolute;
-    inset: auto 16px 16px auto;
-  }
-
-  form:is(.tab-inventory, .tab-features, .tab-spells, .tab-effects) .create-child { display: block; }
 
   /* ---------------------------------- */
   /*  Warnings                          */
