+++ less/v2/character.less
@@ -89,7 +89,6 @@
       .level-badge {
         width: var(--level-size);
         height: var(--level-size);
-        background-image: url("ui/level-badge.webp");
         margin-left: auto;
       }
 
@@ -830,6 +829,17 @@
 
       .pill { padding: .25rem .5rem; }
     }
+
+    /* Item Sections */
+    .inventory-element {
+      gap: 16px;
+      .items-list { gap: 16px; }
+    }
+
+    .effects-element {
+      gap: 16px;
+      .effects-list { gap: 16px; }
+    }
   }
 
   /* ---------------------------------- */
@@ -939,6 +949,7 @@
           height: 23px;
           border-radius: 3px;
           border: 1px solid transparent;
+          text-align: center;
 
           &:hover, &:focus {
             border: 1px solid var(--dnd5e-color-gold);
