+++ less/v2/dark/actors.less
@@ -1,7 +1,5 @@
 :is(.dnd5e-theme-dark, .theme-dark) .dnd5e2.sheet.actor,
 .dnd5e2.sheet.actor.dnd5e-theme-dark {
-  .window-content { background: var(--dnd5e-color-dark-gray) url("../../ui/denim075.png"); }
-
   .ability-score {
     .label { font-weight: bold; }
     .mod { --dnd5e-color-black: var(--dnd5e-color-blue-white); }
