+++ less/v2/dark/apps.less
@@ -17,6 +17,8 @@
     --filigree-border-color: var(--dnd5e-color-blue-gray-3);
     --dnd5e-background-5: var(--dnd5e-highlight-10);
 
+    .window-content { background: var(--dnd5e-color-dark-gray) url("../../ui/denim075.png"); }
+
     select option { background: var(--dnd5e-color-blue-gray-2); }
     .window-resizable-handle { filter: invert(1); }
