+++ less/v2/dark/compendium-browser.less
@@ -2,7 +2,6 @@
   --color-text-primary: var(--dnd5e-color-blue-white);
 
   .window-content {
-    background: var(--dnd5e-color-dark-gray) url("../../ui/denim075.png");
     -webkit-backdrop-filter: none;
     backdrop-filter: none;
