+++ less/v2/dark/forms.less
@@ -1,0 +1,14 @@
+.theme-dark .dnd5e2,
+.dnd5e2.dnd5e-theme-dark {
+  fieldset {
+    border-color: var(--dnd5e-color-blue-gray-3);
+    > legend { color: var(--dnd5e-color-gold); }
+  }
+
+  .form-group {
+    select, input {
+      border: var(--dnd5e-border-gold);
+      background: var(--dnd5e-background-parchment);
+    }
+  }
+}
