+++ less/v2/dark/npc.less
@@ -2,6 +2,8 @@
 .dnd5e2.sheet.actor.npc.dnd5e-theme-dark {
 
   .window-content {
+    background: var(--dnd5e-color-dark-gray) url("../../ui/denim075.png");
+
     &::before {
       content: "";
       position: absolute;
