+++ less/v2/inventory.less
@@ -1,7 +1,7 @@
 .dnd5e2 .inventory-element {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: 8px;
 
   /* ---------------------------------- */
   /*  Common Styles                     */
@@ -205,7 +205,7 @@
   .items-list {
     display: flex;
     flex-direction: column;
-    gap: 1rem;
+    gap: 8px;
     container-type: inline-size;
   }
 
@@ -565,7 +565,7 @@
 
 .dnd5e2 section.currency, .dnd5e2 .currency-style {
   display: flex;
-  gap: 1rem;
+  gap: 8px;
 
   label {
     position: relative;
