+++ less/v2/tooltips.less
@@ -137,16 +137,6 @@
       }
     }
 
-    .info {
-      text-align: left;
-      font-size: var(--font-size-12);
-
-      .materials, .condition {
-        font-size: var(--font-size-11);
-        color: var(--color-text-dark-5);
-      }
-    }
-
     .description {
       text-align: left;
       font-size: var(--font-size-12);
