+++ module/applications/_module.mjs
@@ -7,6 +7,7 @@ export * as components from "./components/_module.mjs";
 export * as dice from "./dice/_module.mjs";
 export * as item from "./item/_module.mjs";
 export * as journal from "./journal/_module.mjs";
+export * as mixins from "./mixins/_module.mjs";
 
 export {default as Accordion} from "./accordion.mjs";
 export {default as Award} from "./award.mjs";
