+++ module/applications/components/checkbox.mjs
@@ -32,7 +32,7 @@ export default class CheckboxElement extends AdoptedStyleSheetMixin(
       height: var(--checkbox-size, 18px);
       aspect-ratio: 1;
     }
-    
+
     :host > div {
       width: 100%;
       height: 100%;
@@ -41,7 +41,7 @@ export default class CheckboxElement extends AdoptedStyleSheetMixin(
       background: var(--checkbox-empty-color, transparent);
       box-sizing: border-box;
     }
-    
+
     :host :is(.checked, .disabled, .indeterminate) {
       display: none;
       height: 100%;
@@ -50,7 +50,7 @@ export default class CheckboxElement extends AdoptedStyleSheetMixin(
       align-items: center;
       justify-content: center;
     }
-    
+
     :host([checked]) .checked { display: flex; }
     :host([indeterminate]) .indeterminate { display: flex; }
     :host([indeterminate]) .checked { display: none; }
@@ -134,6 +134,13 @@ export default class CheckboxElement extends AdoptedStyleSheetMixin(
    */
   set value(value) {
     this._setValue(value);
+  }
+
+  /** @override */
+  _getValue() {
+    // Workaround for FormElementExtended only checking the value property and not the checked property.
+    if ( typeof this._value === "string" ) return this._value;
+    return this.checked;
   }
 
   /* -------------------------------------------- */
