+++ module/applications/components/inventory.mjs
@@ -217,8 +217,8 @@ export default class InventoryElement extends HTMLElement {
       {
         name: "DND5E.ConcentrationBreak",
         icon: '<dnd5e-icon src="systems/dnd5e/icons/svg/break-concentration.svg"></dnd5e-icon>',
-        condition: () => this.actor.concentration?.items.has(item),
-        callback: () => this.actor.endConcentration(item),
+        condition: () => this.actor?.concentration?.items.has(item),
+        callback: () => this.actor?.endConcentration(item),
         group: "state"
       }
     ];
