+++ module/applications/mixins/_module.mjs
@@ -1,0 +1,1 @@
+export {default as DocumentSheetV2Mixin} from "./sheet-v2-mixin.mjs";
