+++ module/data/item/templates/activated-effect.mjs
@@ -126,6 +126,10 @@ export default class ActivatedEffectTemplate extends SystemDataModel {
     // Prepare labels
     this.parent.labels ??= {};
     this.parent.labels.duration = [this.duration.value, CONFIG.DND5E.timePeriods[this.duration.units]].filterJoin(" ");
+    this.parent.labels.concentrationDuration = this.properties?.has("concentration")
+      ? game.i18n.format("DND5E.ConcentrationDuration", {
+          duration: this.parent.labels.duration.toLocaleLowerCase(game.i18n.lang)
+      }) : this.parent.labels.duration;
     this.parent.labels.activation = this.activation.type ? [
       (this.activation.type in CONFIG.DND5E.staticAbilityActivationTypes) ? null : this.activation.cost,
       CONFIG.DND5E.abilityActivationTypes[this.activation.type]
@@ -152,9 +156,9 @@ export default class ActivatedEffectTemplate extends SystemDataModel {
       this.parent.labels.range = range.filterJoin(" ");
     } else this.parent.labels.range = game.i18n.localize("DND5E.None");
 
-    if ( this.recharge ) this.parent.labels.recharge = `${game.i18n.localize("DND5E.Recharge")} [${
-      `${this.recharge.value}${parseInt(this.recharge.value) < 6 ? "+" : ""}`
-    }]`;
+    if ( this.recharge ) this.parent.labels.recharge = `${game.i18n.localize("DND5E.Recharge")} [
+      ${this.recharge.value}${parseInt(this.recharge.value) < 6 ? "+" : ""}
+    ]`;
 
     // Substitute source UUIDs in consumption targets
     if ( !this.parent.isEmbedded ) return;
