+++ templates/actors/npc-sheet-2.hbs
@@ -523,7 +523,7 @@
 
     {{!-- Child Creation --}}
     <button type="button" class="create-child gold-button"
-            aria-label="{{ "SIDEBAR.Create" type=(localize "DOCUMENT.Item") }}">
+            aria-label="{{ localize "SIDEBAR.Create" type=(localize "DOCUMENT.Item") }}">
         <i class="fas fa-plus"></i>
     </button>
