+++ templates/actors/tabs/creature-features.hbs
@@ -39,7 +39,7 @@
                 {{!-- Items --}}
                 <li class="item collapsible {{#unless ctx.isExpanded}}collapsed{{/unless}}" data-item-id="{{ item.id }}"
                     data-entry-id="{{ item.id }}" data-item-name="{{ item.name }}" data-item-sort="{{ item.sort }}"
-                    data-grouped="{{ ctx.group }}" data-ungrouped="{{ ctx.ungroup }}">
+                    data-grouped="{{ ctx.group }}" data-ungrouped="{{ ctx.ungroup }}" data-uuid="{{ item.uuid }}">
 
                     <div class="item-row">
