+++ templates/shared/active-effects2.hbs
@@ -109,6 +109,7 @@
 
     </section>
 
+    {{#if hasConditions}}
     <section class="items-list">
 
         <div class="items-section card">
@@ -140,5 +141,6 @@
         </div>
 
     </section>
+    {{/if}}
 
 </{{ elements.effects }}>
